+++ NEWS.md
@@ -51,9 +51,15 @@ Initial features
     - `abs()` computes Euclidean norm
     - `convex_hull()` computes convex hull (currently just for `coord2d()` objects)
     - `cross_product()` computes a cross product between `Coord3D` vectors
+    - `distance2d()` and `distance3d()` computes Euclidean distances
     - `mean()` computes centroids of coordinates
     - `normal2d()` computes `Coord2D` normals
     - `normal3d()` computes `Coord3D` normals
+
+* `Line2D` R6 class
+
+  + `is_line2d()` tests whether objects are `Line2D` R6 classes
+  + `as_line2d()` costs objects to `Line2D` R6 classes
 
 * `transform2d()` and `transform3d()` create 2D/3D affine transformation matrix S3 classes